Relocate group_create_form
